pyproject.toml
[project]
name = "arhugula"
version = "0.1.0"
description = "Portable Pi-orchestrated agent development environment"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "ruff>=0.11.0",
    "click>=8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "anthropic>=0.47.1",
    "rich>=13.7.0",
    "watchdog>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ty.environment]
python = ".venv"
